extract sponsorship details into own component
