+++ client/src/components/Sponsorships/Sponsorship.tsx
@@ -1,18 +1,16 @@
 import { Fragment, FunctionComponent, useEffect, useState } from "react";
 import {
-    getIdPadded,
     SponsorRecord,
     SponsorshipRecord,
     StudentRecord,
 } from "@nica-angels/shared";
 import { Card, CardBody, CardFooter, CardHeader } from "reactstrap";
-import { KeyValCardBody, LoadingSpinner } from "jack-hermanson-component-lib";
+import { LoadingSpinner } from "jack-hermanson-component-lib";
 import { useStoreState } from "../../store/_store";
 import { StudentClient } from "../../clients/StudentClient";
 import { SponsorClient } from "../../clients/SponsorClient";
 import { Link } from "react-router-dom";
-import { DATE_FORMAT, ID_PADDING } from "../../utils/constants";
-import moment from "moment";
+import { SponsorshipCardBody } from "./SponsorshipCardBody";
 
 interface Props {
     sponsorship: SponsorshipRecord;
@@ -88,45 +86,13 @@ export const Sponsorship: FunctionComponent<Props> = ({
 
     function renderDetails() {
         if (sponsor && student) {
-            const keyValPairs = [
-                {
-                    key: spanish ? "Estudiante" : "Student",
-                    val: (
-                        <Link className="ps-0" to={`/students/${student.id}`}>
-                            {student.firstName} {student.middleName}{" "}
-                            {student.lastName} (#
-                            {getIdPadded(student, ID_PADDING)})
-                        </Link>
-                    ),
-                },
-                {
-                    key: spanish ? "Padrino" : "Sponsor",
-                    val: (
-                        <Link className="ps-0" to={`/sponsors/${sponsor.id}`}>
-                            {sponsor.firstName} {sponsor.lastName}
-                        </Link>
-                    ),
-                },
-                {
-                    key: spanish ? "Pago" : "Payment",
-                    val: `$${sponsorship.payment}`,
-                },
-                {
-                    key: spanish ? "Frecuencia" : "Frequency",
-                    val: `${sponsorship.frequency}/${spanish ? "año" : "year"}`,
-                },
-                {
-                    key: spanish ? "Fecha de Comienza" : "Start Date",
-                    val: moment(sponsorship.startDate).format(DATE_FORMAT),
-                },
-            ];
-            if (sponsorship.endDate) {
-                keyValPairs.push({
-                    key: spanish ? "Fecha de Cierre" : "End Date",
-                    val: moment(sponsorship.endDate).format(DATE_FORMAT),
-                });
-            }
-            return <KeyValCardBody keyValPairs={keyValPairs} />;
+            return (
+                <SponsorshipCardBody
+                    sponsor={sponsor}
+                    student={student}
+                    sponsorship={sponsorship}
+                />
+            );
         }
     }
 };
